add reduction for add1/sub1 for fixnums in cptypes
Reduce (add1 <fixnum>) to (%3%$fxx+ x 1), and add equivalent reductions
for similar functions. This extends  the reductions used when the
argument is a flonum.

The new primitive $fxx+ is similat to fx+, but in case of an overflow it
calls + to get the correct result. This is slightly faster than using +
directly.