correct `powl` that was intended to be `pow`
The intended change was to use `pow` instead of `powl` for x86_64 and
AArch64 macOS. The mangled change caused `powl` to be used even more
broadly, instead. Also, simplify the comment, since extra caveats seem
related to experiments with a similarly incorrect change, and adjust
the test, since results on x86_64 macOS should be reliably accurate.