Add a few improvements for predicates in cptype, in particular for zero? (#960)
* Allow #f and (void) as results of primitive handlers in cptypes

The speciliation handlers can return #f or (void)
in case they don't want to make changes, and then
the default is used.

The use of #f is similar to `define-inline` in cp0,
but (void) allows to use `when`or `cond` to select
the useful cases.

* Improve predicates handling in cptype

When a primitive is marked as a predicate and has a
handler, then try to reduce it as a predicate and then
call the handler for more specific reductions.

* Add rational? to cptypes

* Improve support of zero? in cptypes

In particular, reduce

(if (zero? x) (zero? x) (zero? x))
=>
(if (zero? x) #t #f)
=>
(zero? x)

* Track 0.0 and -0.0 in cptypes

It useful to reduce

(cond
  [(eqv? x 0.0) ...]
  [(eqv? x -0.0) ...]
  [(flzero? x) ...] ; this can be eliminated
  [...])