add `__atomic` foreign convention and improve related flonum unboxing (#976)
The `__atomic` convention promises that a foreign function will not
call back into Scheme, such as through a foreign callable, before it
returns. That promise enables a more direct call to the foreign
function. For example, on my machine, calling

```
(foreign-procedure __atomic "(cs)s_errno" () int))
```

takes about 80% the time of calling

```
(foreign-procedure "(cs)s_errno" () int))
```

When a foreign function has arguments or results that are represented
as flonums, using `__atomic` allows the call to be exposed enough in
intermediate compiler passes so that it can partitcapte in
flonum-unboxing optimizations. In principle, unboxing optimizations
could also apply to non-`__atomic` foreign functions, but the current
calling convention for Scheme functions and the protocol between
Scheme code and the kernel (especially the garbage collector) make
that difficult. Further improvements there are possible, but functions
that fit the `__atomic` constraint can be handled relatively easily.
Probably `__atomic` is wanted for the best performance, anyway, when
unboxing is relevant.

Meanwhile, various problems were interfering with flonum unboxing
around `ftype-ref` and `ftype-set!` on `double` and `float` fields,
and those problems are now fixed.

Closes #921 and uses parts of the example there by @williewillus in
tests.