pb foreign-function interface: refine `union` handling for s390x (#990)
With pb, foreign calls in general use libffi, which does not provide
support for `union` arguments and results. The binding in "c/ffi.c"
approximates `union` handling by constructing a `struct` description
that is likely to be treated the same in the platform's ABI. The
approximation was wrong for s390x in the case of `union`s that contain
only floating-point fields, so add a special case for s390x. There are
surely other platforms that would need similar treatment, but s390x is
relevant because there's no native-code backend for that architecture.