trigger interrupts after long vector operations
For operations that are atomic from the perspective of interrupts but
that may work on large objects, such as `vector-append`, adjust the
trap counter proportional to work done. That way, interrupts are
dispatched in a more timely manner, especially GC interrupts.

The change to "7.ms" is unrelated; wrapping that test with its smaller
list size in a loop could provoke a failure befere these changes.